move client/sync/data.rs to client/arrow_data.rs
