+++ kernel/src/client/sync/mod.rs
@@ -4,7 +4,6 @@ use crate::{EngineInterface, ExpressionHandler, FileSystemClient, JsonHandler, P
 
 use std::sync::Arc;
 
-pub mod data;
 mod fs_client;
 mod get_data;
 pub(crate) mod json;
